# products/_template/api-client/src/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
  baseUrl: `${string}://${string}` | (string & {});
};

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
  /**
   * Detail
   */
  detail?: Array<ValidationError>;
};

/**
 * Health
 */
export type Health = {
  /**
   * Status
   */
  status: string;
};

/**
 * Hello
 */
export type Hello = {
  /**
   * Message
   */
  message: string;
};

/**
 * ItemCreate
 */
export type ItemCreate = {
  /**
   * Description
   */
  description?: string | null;
  /**
   * Title
   */
  title: string;
};

/**
 * ItemRead
 */
export type ItemRead = {
  /**
   * Created At
   */
  created_at: string;
  /**
   * Description
   */
  description: string | null;
  /**
   * Id
   */
  id: string;
  /**
   * Owner Id
   */
  owner_id: string;
  /**
   * Title
   */
  title: string;
  /**
   * Updated At
   */
  updated_at: string;
};

/**
 * ItemUpdate
 */
export type ItemUpdate = {
  /**
   * Description
   */
  description?: string | null;
  /**
   * Title
   */
  title?: string | null;
};

/**
 * MeRead
 */
export type MeRead = {
  /**
   * Email
   */
  email?: string | null;
  /**
   * Id
   */
  id: string;
};

/**
 * Page[ItemRead]
 */
export type PageItemRead = {
  /**
   * Items
   */
  items: Array<ItemRead>;
  /**
   * Next Cursor
   */
  next_cursor?: string | null;
};

/**
 * Problem
 *
 * RFC 9457 problem+json body — declared so it types into the OpenAPI schema.
 */
export type Problem = {
  /**
   * Detail
   */
  detail?: string | null;
  /**
   * Instance
   */
  instance?: string | null;
  /**
   * Status
   */
  status: number;
  /**
   * Title
   */
  title: string;
  /**
   * Type
   */
  type?: string;
};

/**
 * PushTokenCreate
 */
export type PushTokenCreate = {
  /**
   * Device Id
   */
  device_id: string;
  /**
   * Expo Token
   */
  expo_token: string;
};

/**
 * PushTokenRead
 */
export type PushTokenRead = {
  /**
   * Device Id
   */
  device_id: string;
  /**
   * Expo Token
   */
  expo_token: string;
  /**
   * Id
   */
  id: string;
};

/**
 * ValidationError
 */
export type ValidationError = {
  /**
   * Context
   */
  ctx?: {
    [key: string]: unknown;
  };
  /**
   * Input
   */
  input?: unknown;
  /**
   * Location
   */
  loc: Array<string | number>;
  /**
   * Message
   */
  msg: string;
  /**
   * Error Type
   */
  type: string;
};

export type HealthzData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/healthz";
};

export type HealthzResponses = {
  /**
   * Successful Response
   */
  200: Health;
};

export type HealthzResponse = HealthzResponses[keyof HealthzResponses];

export type HelloData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/v1/hello";
};

export type HelloResponses = {
  /**
   * Successful Response
   */
  200: Hello;
};

export type HelloResponse = HelloResponses[keyof HelloResponses];

export type ListItemsData = {
  body?: never;
  path?: never;
  query?: {
    /**
     * Cursor
     */
    cursor?: string | null;
    /**
     * Limit
     */
    limit?: number;
  };
  url: "/v1/items";
};

export type ListItemsErrors = {
  /**
   * Unauthorized
   */
  401: Problem;
  /**
   * Not Found
   */
  404: Problem;
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type ListItemsError = ListItemsErrors[keyof ListItemsErrors];

export type ListItemsResponses = {
  /**
   * Successful Response
   */
  200: PageItemRead;
};

export type ListItemsResponse = ListItemsResponses[keyof ListItemsResponses];

export type CreateItemData = {
  body: ItemCreate;
  path?: never;
  query?: never;
  url: "/v1/items";
};

export type CreateItemErrors = {
  /**
   * Unauthorized
   */
  401: Problem;
  /**
   * Not Found
   */
  404: Problem;
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type CreateItemError = CreateItemErrors[keyof CreateItemErrors];

export type CreateItemResponses = {
  /**
   * Successful Response
   */
  201: ItemRead;
};

export type CreateItemResponse = CreateItemResponses[keyof CreateItemResponses];

export type DeleteItemData = {
  body?: never;
  path: {
    /**
     * Item Id
     */
    item_id: string;
  };
  query?: never;
  url: "/v1/items/{item_id}";
};

export type DeleteItemErrors = {
  /**
   * Unauthorized
   */
  401: Problem;
  /**
   * Not Found
   */
  404: Problem;
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type DeleteItemError = DeleteItemErrors[keyof DeleteItemErrors];

export type DeleteItemResponses = {
  /**
   * Successful Response
   */
  204: void;
};

export type DeleteItemResponse = DeleteItemResponses[keyof DeleteItemResponses];

export type GetItemData = {
  body?: never;
  path: {
    /**
     * Item Id
     */
    item_id: string;
  };
  query?: never;
  url: "/v1/items/{item_id}";
};

export type GetItemErrors = {
  /**
   * Unauthorized
   */
  401: Problem;
  /**
   * Not Found
   */
  404: Problem;
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type GetItemError = GetItemErrors[keyof GetItemErrors];

export type GetItemResponses = {
  /**
   * Successful Response
   */
  200: ItemRead;
};

export type GetItemResponse = GetItemResponses[keyof GetItemResponses];

export type UpdateItemData = {
  body: ItemUpdate;
  path: {
    /**
     * Item Id
     */
    item_id: string;
  };
  query?: never;
  url: "/v1/items/{item_id}";
};

export type UpdateItemErrors = {
  /**
   * Unauthorized
   */
  401: Problem;
  /**
   * Not Found
   */
  404: Problem;
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type UpdateItemError = UpdateItemErrors[keyof UpdateItemErrors];

export type UpdateItemResponses = {
  /**
   * Successful Response
   */
  200: ItemRead;
};

export type UpdateItemResponse = UpdateItemResponses[keyof UpdateItemResponses];

export type MeData = {
  body?: never;
  path?: never;
  query?: never;
  url: "/v1/me";
};

export type MeResponses = {
  /**
   * Successful Response
   */
  200: MeRead;
};

export type MeResponse = MeResponses[keyof MeResponses];

export type RegisterTokenData = {
  body: PushTokenCreate;
  path?: never;
  query?: never;
  url: "/v1/push-tokens";
};

export type RegisterTokenErrors = {
  /**
   * Validation Error
   */
  422: HttpValidationError;
};

export type RegisterTokenError = RegisterTokenErrors[keyof RegisterTokenErrors];

export type RegisterTokenResponses = {
  /**
   * Successful Response
   */
  201: PushTokenRead;
};

export type RegisterTokenResponse = RegisterTokenResponses[keyof RegisterTokenResponses];